backend/requirements.txt
scikit-learn==0.24.1
numpy==1.21.6
opencv-python==4.8.0.76
scipy==1.7.3
joblib==1.3.2
cloudinary==1.34.0
python-dotenv==1.0.0
